requirements.txt
azure-ai-agents
azure-ai-inference
azure-ai-projects
azure-ai-textanalytics
azure-core
azure-identity
azure-keyvault-secrets
certifi
dotenv
flask
matplotlib
openai
pillow
promptflow
pylint
requests
semantic-kernel[azure]
